atoms-muscle/pyproject.toml
[project]
name = "atoms-muscle"
version = "0.1.0"
description = "The Heavy Lifter: Execution Environment for GPU-heavy tasks (Video, Audio, 3D)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "ffmpeg-python",
    "numpy",
    "pydantic",
    "python-multipart",
    "boto3",
    "faster-whisper",
    "torch",
    "open_clip_torch",
    "mcp"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/muscle"]

[tool.uv]
[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx"
]
